+++ pkg/providers/papi/testdata/basic_with_production_activation/import.sh
@@ -1,0 +1,4 @@
+terraform init
+terraform import akamai_edge_hostname.test-edgesuite-net ehn_2867480,test_contract,grp_12345
+terraform import akamai_property.test-edgesuite-net prp_12345,test_contract,grp_12345,LATEST
+terraform import akamai_property_activation.test-edgesuite-net-production prp_12345:PRODUCTION