+++ pkg/providers/papi/testdata/include_single_network/variables.tf
@@ -7,8 +7,3 @@ variable "config_section" {
   type    = string
   default = "test_section"
 }
-
-#variable "env" {
-#  type    = string
-#  default = "staging"
-#}