+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Release Notes
 
+## Version 1.7.0 (July 5, 2023)
+
+### Features/Enhancements
+
+* Migrated to Terraform 1.4.6 version
+
+* PAPI
+  * Added support for `export-property` command with flag `--schema` for properties in frozen formats `v2023-01-05` and `v2023-05-30`. 
+  * Added support for import of `akamai_property_activation` resource.
+  * Added changes in `export-property` command:
+    * Added support for `STAGING` and `PRODUCTION` network configurations for `akamai_property_activation` resource.
+    * Removed support for `var.env` variable.
+    * Added support for `auto_acknowledge_rule_warnings` default value in `akamai_property_activation` resource.
+
+* APPSEC
+  * Added support for `ukraine_geo_control_action` to `modules-security-firewall.tmpl` template for IP/Geo Firewall.
+
 ## Version 1.6.0 (May 31, 2023)
 
 ### Bug fixes